prefer inline tests
